Move admin's user list to specific file
